Read current mountpoints from root fstab
Without a previously generated fstab we have no idea which of the paths
in the root are mountpoints and which are just directories in a mounted
bind mount.

Can we differentiate between a plain directory and a bind mount?
Can we differentiate between a bind mount made by spellbinder and one
that just happens to be in a bind mounted tree, without state?